cleanup and add test skelletons
