--- GifSearcher/GifSearcherTests/GifSample.swift
@@ -1,15 +1,0 @@
-//
-//  GifSample.swift
-//  GifSearcherTests
-//
-//  Created on 6/5/20.
-//
-
-import XCTest
-@testable import GifSearcher
-
-struct GifSample {
-    
-    static let sampleGif = AnimatedGif(url: "http://media0.giphy.com/media/l4Ki5O9v4ZkNiTl7O/giphy.gif",
-                                        urlSmall: "http://media.giphy.com/media/l4Ki5O9v4ZkNiTl7O/giphy-tumblr.gif")
-}
