+++ GifSearcher/GifSearcherTests/SearchGifInteractorTests.swift
@@ -10,24 +10,23 @@ import XCTest
 
 class SearchGifInteractorTests: XCTestCase {
 
+    // MARK - Subject Under Test
+    var sut: SearchInteractor!
+    
     override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
+        super.setUp()
+        setupSearchInteractor()
     }
 
     override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    
     }
 
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    func setupSearchInteractor() {
+        sut = SearchInteractor()
     }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
+    
+    func test_presentGifs_ShouldConfirmPresentWasCalled() {
+        let presenter = GifSear
     }
-
 }
