+++ GifSearcher/GifSearcherTests/SearchGifWorkerTests.swift
@@ -1,0 +1,25 @@
+//
+//  SearchGifWorkerTests.swift
+//  GifSearcherTests
+//
+//  Created by Edgar Barocio on 6/5/20.
+//  Copyright © 2020 Edgar Barocio. All rights reserved.
+//
+
+import XCTest
+
+class SearchGifWorkerTests: XCTestCase {
+
+    override func setUpWithError() throws {
+    
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        
+    }
+
+}
