+++ GifSearcher/GifSearcherTests/SpyClasses.swift
@@ -8,6 +8,12 @@
 import Foundation
 @testable import GifSearcher
 
+struct GifSample {
+    
+    static let sampleGif = AnimatedGif(url: "http://media0.giphy.com/media/l4Ki5O9v4ZkNiTl7O/giphy.gif",
+                                        urlSmall: "http://media.giphy.com/media/l4Ki5O9v4ZkNiTl7O/giphy-tumblr.gif")
+}
+
 class SearchPresenterDisplayLogicSpy: SearchPresentationLogic {
     
     //MARK: - Expectation
